# apps/spark_board_backend/src/swagger/SparkBoard.swagger.yaml
openapi: 3.0.3
info:
  title: SparkBoard
  version: 1.0.0
  contact: {}
servers:
  - url: "{{base_url}}"
paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login
      description: Login
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: albin@mail.com
                password:
                  type: string
                  example: "12345678"
            examples:
              Login:
                value:
                  email: albin@mail.com
                  password: "12345678"
      responses:
        "200":
          description: ""
      security:
        - {}
  /auth/signup:
    post:
      tags:
        - Auth
      summary: Signup
      description: Signup
      operationId: signup
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: albin@mail.com
                name:
                  type: string
                  example: Albin
                password:
                  type: string
                  example: "12345678"
            examples:
              Signup:
                value:
                  email: albin@mail.com
                  name: Albin
                  password: "12345678"
      responses:
        "200":
          description: ""
      security:
        - {}
  /users:
    get:
      tags:
        - User
      summary: List users
      description: List users
      operationId: listUsers
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
  /projects:
    get:
      tags:
        - Projects
      summary: List projects
      description: List projects
      operationId: listProjects
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    post:
      tags:
        - Projects
      summary: Create project
      description: Create project
      operationId: createProject
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: The project
                name:
                  type: string
                  example: Project 1
            examples:
              Create project:
                value:
                  description: The project
                  name: Project 1
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
  /projects/{projectId}:
    delete:
      tags:
        - Projects
      summary: Delete project
      description: Delete project
      operationId: deleteProject
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          example: ff4721f2-047b-4c2d-bce9-14a2c0cbb23b
  /projects/{projectId}/tasks:
    get:
      tags:
        - Tasks
      summary: List tasks
      description: List tasks
      operationId: listTasks
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    post:
      tags:
        - Tasks
      summary: Create task
      description: Create task
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: description
                name:
                  type: string
                  example: The task
                status:
                  type: string
                  example: done
            examples:
              Create task:
                value:
                  description: description
                  name: The task
                  status: done
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          example: 23b9f2df-c215-4d79-bceb-d12e007d0bec
  /projects/{projectId}/tasks/{taskId}:
    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Delete task
      operationId: deleteTask
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    patch:
      tags:
        - Tasks
      summary: Update task
      description: Update task
      operationId: updateTask
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: The description
                name:
                  type: string
                  example: New name
                status:
                  type: string
                  example: todo
            examples:
              Update task:
                value:
                  description: The description
                  name: New name
                  status: todo
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          example: c7a51111-6fe9-41bd-8b17-efb10d911f46
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          example: 4758181a-128c-4e34-b716-2cfb082cfc1a
  /projects/{projectId}/tasks/{taskId}/assign:
    post:
      tags:
        - Tasks
      summary: Assign task
      description: Assign task
      operationId: assignTask
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                staffId:
                  type: string
                  example: cbbe1c2f-ea36-4a1b-b4cc-fa3f17f99af7
            examples:
              Assign task:
                value:
                  staffId: cbbe1c2f-ea36-4a1b-b4cc-fa3f17f99af7
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    delete:
      tags:
        - Tasks
      summary: Remove assign
      description: Remove assign
      operationId: removeAssign
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          example: c7a51111-6fe9-41bd-8b17-efb10d911f46
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          example: 4758181a-128c-4e34-b716-2cfb082cfc1a
  /staff:
    get:
      tags:
        - Staff
      summary: List staff
      description: List staff
      operationId: listStaff
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    post:
      tags:
        - Staff
      summary: Create staff
      description: Create staff
      operationId: createStaff
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: praveen@mail.com
                name:
                  type: string
                  example: Praveen
            examples:
              Create staff:
                value:
                  email: praveen@mail.com
                  name: Praveen
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
  /staff/{staffId}:
    delete:
      tags:
        - Staff
      summary: Delete staff
      description: Delete staff
      operationId: deleteStaff
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    patch:
      tags:
        - Staff
      summary: Update staff
      description: Update staff
      operationId: updateStaff
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Shebin
            examples:
              Update staff:
                value:
                  name: Shebin
      responses:
        "200":
          description: ""
      security:
        - bearerAuth: []
    parameters:
      - name: staffId
        in: path
        required: true
        schema:
          type: string
          example: e98085f6-bc9f-4943-bc9b-a3a089649ae9
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
tags:
  - name: Auth
  - name: User
  - name: Projects
  - name: Tasks
  - name: Staff